requirements.txt
google-generativeai>=0.3.0
click>=8.1.0
colorama>=0.4.6
python-dotenv>=1.0.0
